combine Exporter3Test and ParticipantVersionTest
